requirements.txt
alembic==1.16.5
Authlib==1.3.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.23
python-dotenv==1.0.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.1.2
